Move `gameOver` to the `Frontend` interface.
