+++ winbolo/src/gui/linux/frontend.h
@@ -40,6 +40,7 @@ class LinuxFrontend : public Frontend {
   void killsDeaths(int kills, int deaths) override;
   void manStatus(bool isDead, TURNTYPE angle) override;
   void manClear(void) override;
+  void gameOver(void) override;
 };
 
 }  // namespace bolo
